Backend/pyproject.toml
[tool.poetry]
name = "courseproject"
version = "0.1.0"
description = ""
authors = ["kawaragi <dima.kretinin.05@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
psycopg2 = "^2.9.10"
asyncio = "^3.4.3"
fastapi = "^0.115.5"
yarl = "^1.18.3"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
sqlalchemy-to-pydantic = "^0.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
